// src/app/models/course.ts
export class Course {
    id: number;
    name: string;
    openingYear: string;
    architect: string;
    holes: number;
    pairs: number;
    facilities: string[];
    isClubHouse: boolean;
    isAcademy: boolean;
    website: string;
    local: string;
    county: string;
    district: string;
    latitude: number;
    longitude: number;
}